feat: implemented base form chatting
Yes real time chatting is possible now. Altough it doesn't save messages yet it's fine for now.
Actually no it's not. Now I recommend you create a new issue on user authentication so we could
continue this further. What I mean is right now we need the database for handling messages and to do
that we depend on user and user id.